check if Cucumber is defined before executing tasks
if the cucumber.task exists and the gem isn't installed yet then the
rails gems:install command fails breaking the config.gem process.

the application may be deployed to a server that doesn't have the
cucumber gem yet or pulled from a repository that contains the
cucumber.task but the gem isn't installed locally yet.